Move request entities to api
