+++ src/api/routes/mod.rs
@@ -1,7 +1,6 @@
-pub use ban::process_ban;
+pub use ban::{process_ban, BanRequest};
 pub use config::configuration_handler;
 pub use unban::process_unban;
-
 mod ban;
 mod config;
 mod unban;
